Uber-commit: initial working implementation of the exchange
This commit contains the initial implementation of my stock exchange.
The exchange is a set of 4 services: a database, a queue, an api server,
and an order matcher. The DB and queue are standard services (postgres
and rabbitmq), the api server is a flask server that supports various
CRUD operations on users and orders, and the matcher is a long-running
process that handles matching and "completing" orders.